Explicitly fail when the hiscores API changes
When new rows are added to the hiscores CSV API, old versions of
osrs-json-api can return corrupted data due to incorrect mappings
from rows to skill/boss/etc. To avoid returning corrupted data to
applications using old versions of osrs-json-api, we will treat
this as an explicit failure with an actionable error message.